+++ src/components/RichTextEditor/RichTextEditor.cy.tsx
@@ -8,7 +8,7 @@ import { ON_SAVE_DELAY_IN_MS } from './utils';
 import { EditorActions } from './utils/actions';
 import { value as exampleValue } from './utils/exampleValues';
 import { toPlaintext } from './utils/plaintext';
-import { Position } from './PositioningEditorWrapper';
+import { Position } from './EditorPositioningWrapper';
 
 const RICH_TEXT_EDITOR = '[data-test-id=rich-text-editor]';
 const TOOLBAR = '[data-test-id=toolbar-floating]';
