+++ src/components/RichTextEditor/index.ts
@@ -5,4 +5,4 @@ export * from './utils';
 export { EditorActions } from './utils/actions';
 export { EMPTY_RICH_TEXT_VALUE } from './utils/parseRawValue';
 export * from './serializer/serializeToHtml';
-export { Position } from './PositioningEditorWrapper';
+export { Position } from './EditorPositioningWrapper';
